Fixes to make the ngo lib compile under Solaris 2.7. The biggest change is that the special versions of CascadeBG and ToggleBG are no longer used in Solaris 2.7.  The Solaris CascadeBG still does a few direct color allocations (4 it seems), but this is less of a headache than trying to come up with a working Motif 2.1 version of this routine, since we do not have access to Motif 2.x source.
MegaB needed some more OS-dependent conditional coding.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@5595 10fe2890-8a40-4696-8b1d-7ccbb19c398e

+++ ni/src/lib/ngo/plotapp.c
@@ -1,5 +1,5 @@
 /*
- *      $Id: plotapp.c,v 1.6 1999-09-15 02:23:18 dbrown Exp $
+ *      $Id: plotapp.c,v 1.7 1999-09-28 00:47:32 dbrown Exp $
  */
 /************************************************************************
 *									*
@@ -2044,7 +2044,7 @@ static void MergeObjResDataItems
 				(rprof->qname,qres,obj->qbasename,
 				 obj->class,ditype,rprof->mindims,
 				 rprof->maxdims,
-				 rprof->data < 0 ? NULL : rprof->data,
+				 ((int)rprof->data) < 0 ? NULL : rprof->data,
 				 rprof->required < 0 ? False : rprof->required,
 				 rprof->vis < 0 ? True : rprof->vis,
 				 set_only,
